refactor(server): initiate pandas backend with new signature
No more classes, just pure functions
Steps will follow the same process
